requirements.txt
alembic==1.7.7
certifi==2021.10.8
chardet==3.0.4
click==8.1.3
colorama==0.4.4
filelock==3.6.0
Flask==2.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
huggingface-hub==0.5.1
idna==2.10
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
Mako==1.2.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
psycopg2-binary
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.13.0
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.23.0
sacremoses==0.0.53
six==1.16.0
SQLAlchemy==1.4.36
tokenizers==0.12.1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.18.0
typing-extensions==4.2.0
urllib3==1.25.11
Werkzeug==2.0.0
zipp==3.8.0
